Tags and check for category and tage deletion added if post is associated with them.

+++ app/Http/Controllers/TagsController.php
@@ -96,6 +96,12 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0)
+        {
+            session()->flash('error', 'Tag can not be deleted because it has some posts associated.');
+            return redirect()->back();
+        }
+
         $tag->delete();
         session()->flash('success', 'Tag deleted successfully.');
         return redirect(route('tags.index'));
